services/bot/requirements.txt
aiogram==3.13.1
asyncpg==0.29.0
python-dotenv==1.0.1
openai==1.43.0
SQLAlchemy==2.0.35
pytest==8.3.2
pytest-asyncio==0.23.3
pytest-xdist==3.5.0
httpx==0.26.0
aiohttp==3.9.1
numpy==1.26.4
Pillow==10.4.0

